Moved activity_map_view to common widgets
