+++ lib/presentation/screens/activity/activity_screen_content.dart
@@ -3,11 +3,11 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:movna/presentation/blocs/activity_cubit.dart';
 import 'package:movna/presentation/router/router.dart';
 import 'package:movna/presentation/screens/activity/views/alerts/activity_alerts_view.dart';
-import 'package:movna/presentation/screens/activity/views/map/activity_map_view.dart';
+import 'package:movna/presentation/widgets/activity_map_view.dart';
 
 /// Displays the content of the activity screen.
 ///
-/// Displays the [ActivityMapView] overlayed by eventual
+/// Displays the [ActivityMapView] overlaid by eventual
 /// [ActivityAlert].
 class ActivityScreenContent extends StatelessWidget {
   const ActivityScreenContent({super.key});
