+++ lib/presentation/widgets/activity_map_view.dart
